+++ src/Systemic/Process/Launcher/Unix.php
@@ -31,6 +31,10 @@ class Unix extends Base
         $workingDirectory = $this->workingDirectory !== null ?
             realpath($this->workingDirectory) : null;
 
+        if ($workingDirectory === false) {
+            $workingDirectory = $this->workingDirectory;
+        }
+
         $result = new Result();
 
         $env = $this->prepareEnv();
@@ -43,6 +47,7 @@ class Unix extends Base
         $outputBuffer = $errorBuffer = $input = null;
         $outputPipe = $pipes[1];
         $errorPipe = $pipes[2];
+        $generatorCalled = false;
 
         stream_set_blocking($outputPipe, false);
         stream_set_blocking($errorPipe, false);
@@ -55,13 +60,22 @@ class Unix extends Base
         while (true) {
             $status = proc_get_status($processHandle);
 
+            // Get output & error
             $outputBuffer = $this->readChunk($outputPipe, $this->readChunkSize);
             $errorBuffer = $this->readChunk($errorPipe, $this->readChunkSize);
 
-            if ($this->broker) {
+            // Get input
+            if ($this->inputGenerator) {
+                if (!$generatorCalled) {
+                    $input = ($this->inputGenerator)();
+                    $generatorCalled = true;
+                }
+            } elseif ($this->broker) {
                 $input = $this->broker->read($this->readChunkSize);
             }
 
+
+            // Write output
             if ($outputBuffer !== null) {
                 $result->appendOutput($outputBuffer);
 
@@ -70,6 +84,8 @@ class Unix extends Base
                 }
             }
 
+
+            // Write error
             if ($errorBuffer !== null) {
                 $result->appendError($errorBuffer);
 
@@ -78,9 +94,16 @@ class Unix extends Base
                 }
             }
 
+
+            // Write input
             if ($input !== null) {
                 fwrite($pipes[0], $input);
                 $input = null;
+
+                if ($generatorCalled) {
+                    fclose($pipes[0]);
+                    $pipes[0] = null;
+                }
             }
 
             if (!$status['running'] && $outputBuffer === null && $errorBuffer === null && $input === null) {
@@ -147,7 +170,7 @@ class Unix extends Base
             chdir($cwd);
         }
 
-        return new UnixProcess($pid, $command);
+        return new UnixProcess((int)$pid, $command);
     }
 
     /**
@@ -183,16 +206,16 @@ class Unix extends Base
             $command .= ' '.implode(' ', $temp);
         }
 
-        if ($this->user) {
-            $command = 'sudo -u '.$this->user.' '.$command;
-        }
-
-        if (Systemic::$os->which('script')) {
+        if ($this->decoratable && Systemic::$os->which('script')) {
             if (Systemic::$os->isMac()) {
                 $command = 'script -q /dev/null '.$command;
             } else {
                 $command = 'script -e -q -c "'.$command.'" /dev/null';
             }
+        }
+
+        if ($this->user) {
+            $command = 'sudo -u '.$this->user.' '.$command;
         }
 
         return $command;
@@ -204,7 +227,13 @@ class Unix extends Base
     protected function prepareEnv(): ?array
     {
         if (!in_array(\PHP_SAPI, ['cli', 'phpdbg'])) {
-            return null;
+            $output = $_ENV;
+
+            if (!isset($output['PATH'])) {
+                $output['PATH'] = '/usr/local/sbin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin';
+            }
+
+            return $output;
         }
 
         $env = $_SERVER;